refactor to use CSI sampling to decide whether to initialize promise
