+++ ads/google/a4a/google-data-reporter.js
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-import {EXPERIMENT_ATTRIBUTE, QQID_HEADER} from './utils';
-import {BaseLifecycleReporter, GoogleAdLifecycleReporter} from './performance';
-import {getMode} from '../../../src/mode';
 import {
-  isExperimentOn,
-  toggleExperiment,
-  randomlySelectUnsetExperiments,
-} from '../../../src/experiments';
-
+  EXPERIMENT_ATTRIBUTE,
+  QQID_HEADER,
+  isReportingEnabled,
+} from './utils';
+import {BaseLifecycleReporter, GoogleAdLifecycleReporter} from './performance';
+import {randomlySelectUnsetExperiments} from '../../../src/experiments';
 import {
     parseExperimentIds,
     isInManualExperiment,
@@ -116,29 +114,11 @@ function isInReportableBranch(ampElement, namespace) {
  * @visibleForTesting
  */
 export function getLifecycleReporter(ampElement, namespace, slotId) {
-  // Carve-outs: We only want to enable profiling pingbacks when:
-  //   - The ad is from one of the Google networks (AdSense or Doubleclick).
-  //   - The ad slot is in the A4A-vs-3p amp-ad control branch (either via
-  //     internal, client-side selection or via external, Google Search
-  //     selection).
-  //   - We haven't turned off profiling via the rate controls in
-  //     build-system/global-config/{canary,prod}-config.json
-  // If any of those fail, we use the `BaseLifecycleReporter`, which is a
-  // a no-op (sends no pings).
-  const type = ampElement.element.getAttribute('type');
-  const win = ampElement.win;
-  const experimentName = 'a4aProfilingRate';
-  // In local dev mode, neither the canary nor prod config files is available,
-  // so manually set the profiling rate, for testing/dev.
-  if (getMode().localDev) {
-    toggleExperiment(win, experimentName, true, true);
-  }
-  randomlySelectUnsetExperiments(win, PROFILING_BRANCHES);
-  if ((type == 'doubleclick' || type == 'adsense') &&
-      isInReportableBranch(ampElement, namespace) &&
-      isExperimentOn(win, experimentName)) {
-    return new GoogleAdLifecycleReporter(win, ampElement.element, namespace,
-        Number(slotId));
+  randomlySelectUnsetExperiments(ampElement.win, PROFILING_BRANCHES);
+  if (isReportingEnabled(ampElement) &&
+      isInReportableBranch(ampElement, namespace)) {
+    return new GoogleAdLifecycleReporter(ampElement.win, ampElement.element,
+      namespace, Number(slotId));
   } else {
     return new BaseLifecycleReporter();
   }
@@ -204,4 +184,3 @@ export function setGoogleLifecycleVarsFromHeaders(headers, reporter) {
   pingParameters[renderingMethodKey] = headers.get(renderingMethodHeader);
   reporter.setPingParameters(pingParameters);
 }
-
