+++ ads/google/a4a/utils.js
@@ -25,6 +25,10 @@ import {isProxyOrigin} from '../../../src/url';
 import {viewerForDoc} from '../../../src/services';
 import {base64UrlDecodeToBytes} from '../../../src/utils/base64';
 import {domFingerprint} from '../../../src/utils/dom-fingerprint';
+import {
+  isExperimentOn,
+  toggleExperiment,
+} from '../../../src/experiments';
 
 /** @const {string} */
 const AMP_SIGNATURE_HEADER = 'X-AmpAdSignature';
@@ -90,6 +94,33 @@ export function isGoogleAdsA4AValidEnvironment(win) {
   // mode, regardless of origin path.
   return supportsNativeCrypto &&
       (isProxyOrigin(win.location) || getMode().localDev || getMode().test);
+}
+
+/**
+ * @param {!AMP.BaseElement} ampElement The element on whose lifecycle this
+ *    reporter will be reporting.
+ * @return {boolean} whether reporting is enabled for this element
+ */
+export function isReportingEnabled(ampElement) {
+  // Carve-outs: We only want to enable profiling pingbacks when:
+  //   - The ad is from one of the Google networks (AdSense or Doubleclick).
+  //   - The ad slot is in the A4A-vs-3p amp-ad control branch (either via
+  //     internal, client-side selection or via external, Google Search
+  //     selection).
+  //   - We haven't turned off profiling via the rate controls in
+  //     build-system/global-config/{canary,prod}-config.json
+  // If any of those fail, we use the `BaseLifecycleReporter`, which is a
+  // a no-op (sends no pings).
+  const type = ampElement.element.getAttribute('type');
+  const win = ampElement.win;
+  const experimentName = 'a4aProfilingRate';
+  // In local dev mode, neither the canary nor prod config files is available,
+  // so manually set the profiling rate, for testing/dev.
+  if (getMode().localDev) {
+    toggleExperiment(win, experimentName, true, true);
+  }
+  return (type == 'doubleclick' || type == 'adsense') &&
+      isExperimentOn(win, experimentName);
 }
 
 /**
